Book repository and package refactor
