+++ app/src/main/res/layout/activity_book_list.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".BookListActivity">
+        tools:context=".ui.book.BookListActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="0dp"
