+++ app/src/main/res/layout/add_book_layout.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        tools:context=".BookListActivity" android:paddingStart="20dp" android:paddingEnd="20dp"
+        tools:context=".ui.book.BookListActivity" android:paddingStart="20dp" android:paddingEnd="20dp"
         android:paddingTop="50dp"
         android:paddingBottom="50dp">
     <EditText
